Add start trigger -> Enter Add move and shot only after start game. Stop player move/shot after death. Add GameOver screen after death Add collision : Player->Alien, Alien->Bullet

+++ src/main/java/pl/pacinho/alienkiller/logic/AlienMoveThread.java
@@ -26,7 +26,7 @@ public class AlienMoveThread extends Thread implements ActionListener {
 
     @Override
     public void actionPerformed(ActionEvent e) {
-        alien.setX(alien.getX() - 10);
+        alien.setX((int)alien.getX() - 10);
         if (alien.getX() == 0) {
             timer.stop();
             currentThread().interrupt();
